feat: 实现多字体 fallback 渲染与排版优化
新增字体栈系统，支持按文本字符自动选择最佳字体；重构字号处理与文本排版逻辑，修复多项细节问题。

+++ src/core/elements.rs
@@ -69,7 +69,7 @@ pub enum PlotElement {
         x: f64,
         y: f64,
         text: String,
-        fontsize: i32,
+        fontsize: f64,
         color: RgbColor,
         font_family: Option<String>,
     },
